Add VagaOcupada method and update deletion logic
- Introduced `VagaOcupada` method in `IVagaRepository` to check if a "vaga" is occupied.
- Implemented the method in `VagaRepository` to retrieve and assess occupancy status.
- Modified `VagaController` to prevent deletion of occupied "vagas", returning a `BadRequest` if deletion is attempted on an occupied slot.

+++ src/ParkingOnline.WebApi/Controllers/VagaController.cs
@@ -50,6 +50,13 @@ public class VagaController(IVagaRepository vagaRepository) : ControllerBase
                 return NotFound($"Não há vaga cadastrada com o id {id}.");
             }
 
+            var vagaOcupada = await vagaRepository.VagaOcupada(id);
+
+            if (vagaOcupada)
+            {
+                return BadRequest("Não é possível deletar uma vaga que está ocupada.");
+            }
+
             await vagaRepository.DeleteVagaAsync(id);
 
             return NoContent();
